+++ demo-site/src/main/java/app/DemoSiteApp.java
@@ -11,6 +11,8 @@ public class DemoSiteApp extends App {
     protected void initialize() {
         load(new SystemModule("sys.properties"));
 
+        redis().host("localhost");
+
         load(new WebModule());
     }
 }